Move AV related actions to AVEx and refactor tests for stability
There were a lot of issues with GUID collisions that didn't make much sense. I suspect some kind of failures were causing the blueprint cache to not get cleared.

Now all tests refer to the same test data (using a static import for readability), which is explicitly set to null as well as cleared from the cache after each test. If nothing else, this should make it easier to debug issues if they crop up again.