Process all ActionBuilder templates
Reverted back to the idea of just taking the class "as is" to generate the output result. The only differences between the output and template:
* Different namespace (this way compilation/intellisense detects problems w/ the templates)
* Implements attributes (these are part of the code gen lib--I could include them in the final output but it doesn't seem useful. If someone convinces me it's easy enough to migrate the attr to the output.)
* Any BlueprintCoreGen code is removed (attributes, usings)

Later on I'll add a [Replace] comment annotation to allow replacing things as needed, e.g. internal references.

Next step is to fill in the unimplemented action types!