Define new IField and IFieldEnumerable interfaces w/ basic implementations
The new definitions attempt to make the scope of Field more clear. Validation is being moved from Field to Method so Field doesn't need to be aware of where it's being used. Instead Field exposes ShouldValidate for use in Method.

Everything is shortened a little too:
-Merged type name overrides into a single check by type
-Added nullable primitive handling

Temporarily the existing code has been renamed to some variation of Old for naming conflicts.