Update FieldParameter to dynamically use internal values
The interface exposed to MethodFactory doesn't change, but the internal implementation is more fine grained to expose better manual overrides. With the IFieldParameter interface doing something like adjusting the parameter name requires overriding the comment, assignment, etc.

This also moves to a model where all input values default to null which is treated as not-set, i.e. if a parameter is null then it will not have any impact on the object.

Additionally the new definition supports fallback assignments for null values. This can be used to implement fallback to things like Constants.Empty as well as ensuring that Lists and Arrays are always initialized (and maybe Dictionary).